requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-rest-knox==5.0.2

PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pytesseract==0.3.13
pdfminer.six==20191110
opencv-python==4.11.0.86
Pillow==11.2.1

pandas==2.2.3
numpy
requests==2.32.3
python-dotenv==1.1.0
PyYAML==6.0.2
ddgs
pytest

spacy==3.8.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

openai==1.73.0
sentence-transformers==4.1.0
langgraph
langchain==0.3.23
langchain-openai==0.3.13
langchain-qdrant==0.2.0
langchain-core==0.3.52
langchain-community==0.3.21
langchain-text-splitters==0.3.8
langchain-ollama==0.2.3
qdrant-client==1.13.3
fastembed

moviepy==1.0.3
SpeechRecognition
pydub
openai-whisper
psycopg2-binary

comtypes
rank_bm25
whisper
nltk
bs4
drf_yasg
tiktoken
langdetect

chardet
rarfile
py7zr
python-magic
pocketsphinx
docx2txt
backoff
xlrd
extract-msg